Something odd is happening here
The SHA generated by nixos appears to be different than the hash
generated by nix darwin in this flake. This could be a side effect of
the system version change (from unstable to stable). For now, I have
more faith in the "correctness" of nixos's SHA and will deal with nix
darwin later.